+++ src/test/java/picocli/CommandLineHelpTest.java
@@ -3068,9 +3068,28 @@ public class CommandLineHelpTest {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
         app.usage(new PrintStream(baos));
 
-        String expected = "Usage: app\n" +
-                          "Commands:\n" +
-                          "  foo  This is a foo sub-command\n";
+        String expected = format("" +
+                "Usage: app%n" +
+                "Commands:%n" +
+                "  foo  This is a foo sub-command%n");
+        assertEquals(expected, new String(baos.toByteArray(), "UTF-8"));
+    }
+
+    @Test
+    public void testUsage_NoHeaderIfAllSubcommandHidden() throws Exception {
+        @Command(name = "foo", description = "This is a foo sub-command", hidden = true) class Foo { }
+        @Command(name = "bar", description = "This is a foo sub-command", hidden = true) class Bar { }
+        @Command(name = "app", abbreviateSynopsis = true) class App { }
+
+        CommandLine app = new CommandLine(new App())
+                .addSubcommand("foo", new Foo())
+                .addSubcommand("bar", new Bar());
+
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        app.usage(new PrintStream(baos));
+
+        String expected = format("" +
+                "Usage: app%n");
         assertEquals(expected, new String(baos.toByteArray(), "UTF-8"));
     }
 }